--- src/main/kotlin/com/marsrover/NasaMarsRover/domain/entities/Plateau.kt
@@ -1,9 +1,0 @@
-package com.marsrover.NasaMarsRover.domain.entities
-
-import com.marsrover.NasaMarsRover.domain.interfaces.IPlateauInterface
-import com.marsrover.NasaMarsRover.domain.value_objects.Size
-
-data class Plateau(override val size: Size): IPlateauInterface {
-
-}
-
