--- src/main/kotlin/com/marsrover/NasaMarsRover/domain/interfaces/IPlateauInterface.kt
@@ -1,8 +1,0 @@
-package com.marsrover.NasaMarsRover.domain.interfaces
-
-import com.marsrover.NasaMarsRover.domain.value_objects.Size
-
-interface IPlateauInterface {
-    val size: Size
-}
-
