+++ src/main/kotlin/com/marsrover/NasaMarsRover/platuae/entities/Plateau.kt
@@ -1,0 +1,9 @@
+package com.marsrover.NasaMarsRover.platuae.entities
+
+import com.marsrover.NasaMarsRover.platuae.interfaces.IPlateauInterface
+import com.marsrover.NasaMarsRover.platuae.value_objects.Size
+
+data class Plateau(override val size: Size): IPlateauInterface {
+
+}
+
