+++ src/main/kotlin/com/marsrover/NasaMarsRover/platuae/interfaces/IPlateauInterface.kt
@@ -1,0 +1,8 @@
+package com.marsrover.NasaMarsRover.platuae.interfaces
+
+import com.marsrover.NasaMarsRover.platuae.value_objects.Size
+
+interface IPlateauInterface {
+    val size: Size
+}
+
